Pull latest reviews from entry form
